Silence configuration warning
There is a warning in the debug console when starting the extension
(via pressing F5 when the extension directory is open in vscode).
The warning is related to accessing python.pythonPath (a scoped
resource) without providing a resource or null.

Providing null silences the warning. It is probably harmless, but
fixing it cannot hurt