Ensure relative pythonPath is not a folder
`${workspaceRoot}/${pythonPath}` can point at a directory in some cases. As a directory cannot be run, we shouldn't override the property.

As an example, this issue actually affects this repo - with the default value of `pythonPath` - `"python"`, it will find the `python` folder (containing `preview.py`), and return the path to it. This causes any execs using the incorrect path to fail