Make "tab" keybinding compatible with other extensions
I was using another extension (GitHub CoPilot) and I was finding that the tab-to-complete wasn't working with it. Similar to the diagnosis at https://github.com/yzhang-gh/vscode-markdown/issues/1011, the keybinding for this extension was taking precedence, eating the tab before CoPilot could get it. Therefore, this PR makes the keybinding more strict in when it accepts a "tab" event.

Tested on my machine, it fixes the problem as described in that issue. Not sure if this will affect other aspects of this extension?